requirement.txt
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
django-otp 
pyotp
sendgrid